requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
future==0.18.2
pycparser==2.20
pymongo==3.10.1
python-dotenv==0.13.0
python-telegram-bot==11.1.0
six==1.14.0
